add utility for parsing server groups (#551)
For some of the IPC fields we may need to fill them in
using data passed via headers. To reduce the number of
headers the plan is to just send the server group and
rely on the [naming conventions] to extract the app and
cluster. Frigga is not used for the parsing as this
could be used in a hot path and it is rather inefficient.

Comparing throughput of this approach (spectator) with
the Frigga library (frigga):

```
Benchmark      Mode  Cnt        Score       Error   Units
spectator     thrpt   10  3216353.588 ± 87966.348   ops/s
frigga        thrpt   10    11745.242 ±  2367.298   ops/s
```

Comparing allocations:

```
Benchmark      Mode  Cnt        Score       Error   Units
spectator     thrpt   10      784.171 ±     0.006    B/op
frigga        thrpt   10   124987.101 ±    11.368    B/op
```

[naming conventions]: https://github.com/Netflix/iep/blob/master/iep-nflxenv/README.md#server-group-settings