support bounded threshold for percentile types
The percentile approximation types can be quite expensive
due to the number of buckets needed to get a good accuracy
for an arbitrary range. This change adds a builder for the
types that allows a threshold and accuracy to be specified.

The threshold will cap the max value for the computed data
distribution. The accuracy will be used to restrict the set
of buckets to a more reasonable number leading up to the
threshold. This is based on the common operational use case
of understanding whether the latency distribution is
approaching failure more so than actually needing an
accurate percentile. Percentiles that are closer to the
threhsold will still be accurate, but those further away
may be inflated.