use AtlasRegistry if atlas-client is new enough (#559)
To enable a smooth transition from ServoRegistry to
AtlasRegistry for internal users, this change detects
whether a new class in the internal client is present
and sets the registry accordingly. This should allow
us to release the newer version of Spectator and have
some users pick it up before the new version of the
client is released.