update CI to build on jdk10 (#564)
JDK11 is expected to be the next long term stable release
that we'll use internally at Netflix. However, it is not yet
supported by gradle, spotbugs, etc. This bumps are CI
validation to test on JDK8 and JDK10. JDK9 has been
dropped as there isn't any reason to use it rather than
JDK10 at this point.