support bounded range for percentile types (#555)
The percentile approximation types can be quite expensive
due to the number of buckets needed to get a good accuracy
for an arbitrary range. This change adds a builder for the
types that allows a range to be specified.

The range will cap the min and max value for the computed
data distribution. This is based on the common operational
use case of understanding whether the latency distribution
is important once it is large enough and when it approaches
failure. This can often be determined based on timeouts for
clients.