initial version of servlet filter for ipc metrics (#598)
Simple servlet filter that provides the common IPC metrics
for any filtered requests. Endpoint mapping should work as
long as the `getServletPath` method is set correctly for the
filter. In the case of guice we have a hack to detect the
path mapping correctly even though it isn't being set correctly
by guice.