give ObjectName props priority over attributes
There was a report that the `id` for Tomcat thread pool
metrics changed when moving to a newer client version
that used `spectator-ext-jvm` instead of `servo-tomcat`.
The reason is that the key `name` is both in the ObjectName
and the attributes. The attributes were overwriting the
ObjectName property. For now we are just giving precedence
to the ObjectName properties. Longer term we can look at a
better way to distinguish between them so the user can
specify.