improve churn resistance for mostFrequent limiter
Updates the most frequent cardinality limiter to better
handle cases where there is a long tail with a lot of
churn. It should now detect a baseline frequency that
is higher than the frequncy of values with a lot of
churn over time. This may limit the number of values
emitted to be smaller than the limit specified by the
user.