helper for more efficient pattern matching (#651)
There are a number of use-cases for Spectator and Atlas where
we have to support basic regular expressions. This provides
a simple implementation that can handle the aspects we need
in a more efficient manner than the Pattern class in the JDK.
In particular, since we only care if there is a match, there
is no need to keep track of the groups and collect the matching
sub-strings. This allows for some additional optimizations
and means most allocations can be avoided while performing
a check.

With JDK11 we see considerably different performance for
matching regular expressions compared to JDK8. In some cases
it got much better due to some fixes in the JDK11. In others
it seems to have improved due to the changes in how string
encoding work. In particular case insensitive matching is
much faster for a latin1 string. However, in practice on some
of our use-cases we see it can get bogged down resetting the
matcher when running on JDK11. This matcher does not support
capturing groups so we can avoid the state that needs to be
maintained by the JDK.

See benchmark for some examples with common regex styles that
get used with metric data. It also compares with RE2/J which
avoids the worst case behavior of the JDK matcher, but performs
quite poorly for many common patterns we care about.

Looking at the timing on some real data and queries to see
the improvements in overall matching time:

|                        | **Java** | **Custom** |
|-----------------------:|---------:|-----------:|
| **JDK 8**              |   408.0s |     102.4s |
| **JDK 11**             |   312.6s |      56.6s |
| **JDK 11 Ignore Case** |   310.5s |     156.7s |

The slow down when ignoring the case is because the String ops
that ignore the case seem to be quite a bit slower than the
normal operations like `startsWith` and `indexOf`.