Make decimal buckets use 4 digits as needed
Previous to this change using something like
`BucketFunctions.decimal(2_000_000)` produced the following buckets:
`[negative, 0M, 0M, 1M, 2M]`

This change puts it in line with the other functions that will use 4
digits for the bucket boundaries when needed. This means we will
generate buckets: `[negative, 0250k, 0500k, 1000k]` for the above call.