add trace logging of raw payload
For the main publish payload this will be JSON instead
of the binary format over the wire.

Also explicitly adds the URI to failure warning logs as
a number of the default network exceptions do not include
it. The URI and exception is available via the IPC access
logging, but that is debug level and not on by default for
most apps.