add rollup limiter (#679)
This is a simple limiter that limits to a single pre-defined
value if the cardinality of the input exceeds a fixed size.
Useful for values with a fairly uniform frequency that do not
yield decent results with a most frequent limiter. Also used
as a building block for automatic-rollup policies to protect
the backends.