IPC integration for AWS SDK v2 (#689)
Updates `spectator-ext-aws2` to work with the 2.x release
versions. The old metrics collector interfaces are no longer
supported, but there is a generic ExecutionInterceptor that
can be used to get the required information. Instead of
custom metrics, this uses `spectator-ext-ipc` so the it will
report the same data as other Netflix service calls.