reduce allocs for lookups of Percentile types
For percentile approximations that are dynamically looked
up based on the the id, a new instance of the composite
was being created each time. The array used for keeping
track of the counters is about 1-2K (depending on pointer
size) leading to a lot of allocations if this was done in
a hotspot in the code.

This change will cache the results as part of the registry
state. JMH benchmarks show no degradation in the overhead
to record a value and much lower allocation rate:

**Before**

```
defaultTimerGet             16.012 ±       0.001    B/op
defaultTimerReuse            0.005 ±       0.001    B/op
percentileTimerBuilder    1424.163 ±       0.021    B/op
percentileTimerGet        1376.149 ±       0.007    B/op
percentileTimerReuse         0.012 ±       0.001    B/op
```

**After**

```
defaultTimerGet             16.012 ±       0.001    B/op
defaultTimerReuse            0.005 ±       0.001    B/op
percentileTimerBuilder      48.018 ±       0.001    B/op
percentileTimerGet           0.015 ±       0.001    B/op
percentileTimerReuse         0.013 ±       0.001    B/op
```