use double for total on AtlasTimer
In some rare cases we have seen enough activity within a
single interval to overflow the `long` value. Switches it
to a `double` to further reduce the probability of
overflowing the type.

The `totalTime` method is part of the public API and returns
a `long`. This cannot be changed at this time and is mostly
used for unit tests. For now it will just cast the `double`
to a `long`. We'll revisit in the future if it becomes a
problem.