avoid using provider for Registry binding (#699)
In some internal use-cases where this gets used with Spring
via a Guice bridge, the provider can cause problems and avoid
the detection used to prevent interference with the Spring
bindings. While the underlying problem is some bug with the
Guice bridge used in the Spring apps, this change should help
reduce the probability that our library is in the middle of it
when users hit problems.