Revert "avoid using provider for Registry binding (#699)" (#703)
This reverts commit 90b78aea00d0837b9a9332a853dfdc62135fdfd9.

Due to a bug in governator that can cause a memory leak
(see Netflix/governator#392) we are going to revert this
change.

Note, if we do roll forward there is also an intialization
order issue because the SpectatorContext may not be set until
much later. That issue can be worked around by setting it to
the global registry when configured.