add rollup policy for AtlasRegistry
Adds a setting for a RollupPolicy that will get applied to
the set of measurements before sending to Atlas. This can
be used to perform automatic cardinality limiting or other
sanity checks prior to sending to the storage layer.

The policy is not applied for the streaming path so that
full data can be accessed on demand when needed.