add index for matching against a set of queries (#723)
Adds a `QueryIndex` class that can be used to quickly check
and `Id` against a potentially large set of queries that are
known in advance.

This is similar to the structure in the Atlas backend, but
exploits the current structure of the Spectator `Id`. The tree
is setup to check the name first and then the tags in order.
To compensate for regex and other query types that would require
a linear scan of events a simple LFU cache is used. The scan
is only needed for the values that are not in the cache or after
the index is modified at that level.

Tests with real data shows it performs pretty well and allows us
to comfortably check ~450k distinct ids against ~50k alerting
queries in a fraction of a second.