support multiple step sizes for LWC
Updates the LWC integration to support subscribing to
expressions with various step sizes. The meters will
collect data with a base step size. LWC subscriptions
can use any even multiple of that. The data to publish
to Atlas storage will be generated by consolidating the
measurements collected with the base LWC step. The step
size used for publishing must be greater than or equal
and an even multiple of the base LWC step.