avoid exceptions if map is modified concurrently
This change special cases `ConcurrentMap` to ensure that the
code will not throw if there is a concurrent modification
while it is being added to the list. A concurrent modification
should be considered a bug in the caller, but it should not
propagate and cause a bigger problem for the app. Also it
is not possible to detect all cases, only ones that fail in
a noticeable way.

The docs for `Id` have also been updated to clarify the
expectation for the user. We may revisit later to improve
the error reporting for the user, but that requires more
thought.

Fixes #733.