support concurrent posts of metric batches
Before we limited it to a single thread to bound the
overhead. This change will allow a number of concurrent
sends to help for some apps such as the aggregation
service that publish a lot of metrics.

The implementation should allow for substituting the
Java 11 HttpClient in the future either when we drop
Java 8 support or if we decide to build as an MR jar.