add default implementations of new methods to Id (#749)
Though not recommended, in some cases users are implementing
the `Id` interface rather than using `registry.createId`. This
change adds some default implementations to avoid breaking
older code for changes introduced in #722. It also updates the
base registry implementation to normalize the ids to force them
to be a proper implementation created by the registry and avoid
confusing behavior.