cache normalized ids (#754)
In some cases users are using a custom Id implementation to
get better performance at the call site by delaying the dedup
and sorting. In #749 we normalize the Id to avoid any problems
with correctness, but this can lead to a performance penalty
if those users are creating customs ids in a tight loop.

This change adds a simple cache of the normalized id to make
the lookup much cheaper. This avoids most of the overhead on
the sample application that was tested.