split measurement collection from update (#755)
When polling meters, split out the steps of getting all of
the measurements and updating the consolidators and evaluator.
In cases with a lot of meters the second step may take a while
and can potentially cause some meters to be sampled after
crossing a step boundary. This change should make that unlikely
and help confirm the theory is correct.