adjust initial polling delay (#756)
Change it to be close to the start of the interval. With
smaller step sizes this gives more time to publish data
when there are a lot of metrics on an instance. It also
helps get the data to the server more quickly after an
it is complete.

The main down side we'll need to watch out for is that now
the metrics collection and publishing will be more aligned
across all instances of a cluster. From initial testing it
is not expected to cause problems.