avoid storing ids for CompositeRegistry (#764)
If instrumented properly, the number of distinct ids should
not grow unbounded for the life of the process. However, if
users are not disciplined enough with their tagging, this can
occur. With the previous implementation of CompositeRegistry
it would keep track of all of the ids it saw so it could
update the mapping if the registries ever changed. This would
lead to the set of ids growing and eventually using up all
of the heap.

With this change, the composite registry no longer stores the
ids, it just returns a new copy of the SwapMeter instance. A
version is used to indicate if a change has occured that would
require it to perform a new lookup.

This will result in additional allocations if the same id is
looked up repeatedly. Before it would have just returned the
meter instance from the map.