immediately stop polling in PolledMeter.remove
Before it would remove the meter from the state map and
then wait for the object to get garbage collected. In
some cases this could take a while and if a new meter
was registered with the same id, then the value could
get overwritten. Now the polling will get stopped as
part of the remove.