Update ipc.md
I'd like to update the description of the ipc.failure.injected slightly since it can also happen when injecting latency in addition to failure. I've also updated the valid values.

In our environment currently, there is also a "false", but I'm hoping to track that down and clean it up.