shade jackson for spectator-reg-atlas
Internally there are some apps that cannot move beyond
guava 19. When spectator updated to jackson 2.10.0 an
alignment rule forced all jackson subprojects to that
version include jackson-datatype-guava which as of
2.10.0 now dependes on guava 20.

To try and minimize the chances we will be in the middle
of such a mess in the future, this change shades the
jackson usage. The dependency has been relocated, inlined,
and POM file has those dependencies removed.