avoid NPE if registry is shutdown without starting
If the application has a bug in its initialization, then
we sometimes see the stop method get called without the
registry having been started. This would result in an NPE
getting logged because the sender pool was not setup. With
this change the send will get ignored if the pool is not
setup. Should avoid support requests from users asking
about the exception.