clarify behavior for polling static methods
Adds additional helpers for use with static methods to make
that use-case more clear. There is nothing prevent other
usage, but hopefully the naming makes it clear and users will
think about it first.

For existing methods, the comments have been updated to discuss
the behavior of null explicitly and use of null will now get
emitted as a warning.

Fixes #819.