atlas: add support for rule based rollup
This is similar to the aggregate config used on the legacy
internal client. The rollup policy has been refactored so
that it can operate on the common tags as well as user tags.

This is a breaking change if anyone was already using the
RollupPolicy interface, but couldn't find any examples and
it was considered an experiment before. Unfortunately that
was not clearly marked.