fix promotion and live data size for new GCs (#847)
ZGC and Shenandoah are not generational and have a
single reported memory pool. This pool name will be used
for both the young gen to compute the allocation rate as
well as the old gen to compute the promotion rate.

Live data size also checks for entries with all zeroes
that sometimes get reported causing bogus values to show
up for the user.