support negative values for max gauge
This change makes the initial value for a max gauge `NaN`
rather than `0`. As a result negative values will be
treated as being larger and can be used with the max gauge
types.

The benefit is that when using an aggregator normal gauges
flow into a max guage on the remote service. So with this
change those gauges should behave more consistently with
expectations.

Note, this does mean that if someone was using a max gauge
directly via this library, it will now init to `NaN` on
each step interval rather than `0.0`. That is not a common
use-case so it is not a big concern. The local behavior
for the max on timers and distribution summaries has not
changed which is the common use-case.