ipc: add strict validation mode (#886)
Adds a flag that can be used to enable strict validation
for IPC metrics. The goal is that this can be used in unit
tests for integrations to promote consistency.

For now the strict validation just ensures that the only
dimensions on an id are those that are allowed by the spec.
Additional checks can get added overtime which may break
tests for some integrations, but that should help catch
issues and drive compliance going forward.