add a faster QueryIndex.remove
Adds a remove method to the query index that works in the
same way as the add method. The query used when adding the
value must be specified which allows us to avoid searching
the entire index for the value to remove. Testing with a
randomly generated index with 100k queries, the new remove
variant averages 6us per call vs 25ms for the existing
remove variant.