only parse exprs if needed
Before the expression string would get parsed whenever
it was set on the Subscription object. This can be slow
in particular for exprs with complex regex where the
pattern matching optimizer will need to make many passes.

In most cases, after the first fetch the set of expressions
is fairly consistent and only a few new expressions will
need to be parsed. This change makes the parsing lazy.

The other result of this change is that a bad expression
will not be noticed until it is accessed. The evaluator
has been updated with some additional error handling and
logging.