helper to run task using PolledMeter pool (#957)
Adds a `PolledMeter.poll` method that can be used to run
an arbitrary `Runnable` task using the default thread pool
for polled meters. The main use-case is for being able to
update multiple meters each time the task is run. In this
case, the user would just use the registry to make any
updates directly.

The advantage of this method over just having the user
scheduling it themselves is it avoids having additional
polling threads running for simple use-cases. It is not
connected to the lifetime of an object, so the caller must
keep track of and use the returned future to cancel if
the polling should be stopped.