improve indexing for :has and :re
Two improvements to make it more efficient to check
incoming ids:

1. The :has queries are no longer lumped into the other
   checks as all that needs to be verified is that the
   key is present.
2. Uses a prefix tree for the other checks. The pattern
   matcher for :re queries can be used to extract an
   exact match prefix if possible for a regex. This makes
   it much faster to perform the matching if there are a
   lot of prefix regex queries. Looking at the current
   set of expressions internally, this is quite common.

Running a test with an index of ~341k actual expressions
and a real set of ids for incoming datapoints, the indexing
improvements significantly increase the throughput in terms
of datapoints per second checked via the index.

| Test                         | Throughput |
|------------------------------|------------|
| Baseline                     |     71,719 |
| After `:in` expansion (#960) |     97,209 |
| After this change            |    184,856 |

This change also drops an inefficient remove method from the
index to simplify ongoing maintenance.