support expanding OR clauses in regex (#970)
Add helper to PatternMatcher for expanding OR clauses
to a set where the original pattern should match if
at least one of the expanded patterns match. This can
be a useful first step for other processing such as
indexing the query or rewriting for data stores that
support a more limited pattern matching.