support simplifying regex to expression
Adds method to `PatternMatcher` that attempts to rewrite
the pattern to a set of simple pattern matches that can
be combined with AND, OR, and NOT to have the same matching
behavior as the original regex pattern. This can be useful
when working with data stores that have more restricted
pattern matching support such as RE2.