Special case merging ArrayTagSets and TagLists (#986)
Special case merging an ArrayTagSet with another to avoid the
overhead of sorting, deduping, and copying. Since an ArrayTagSet
is immutable, already sorted, and deduped, we can skip directly
to merging..

Additionally, add handling for merging of an ArrayTagSet with a
TagList, where we can avoid some extra copying and allocating of
Tags.