Avoid allocating extra space when overriding tag
Change ArrayTagSet::add to avoid allocating extra space when an existing tag
is being overridden, and to avoid creating a new ArrayTagSet at all if the
supplied key/value pair already exists in the target ArrayTagSet.

Additionally, switch to System.arraycopy instead of manually copying to the
newly created array.