add jdk19 to CI (#1001)
Updates the actions to add JDKs from newest to oldest
so that JAVA_HOME at the end will be the older version.
That version should be used by gradle.

Remove special case checks for guice dependencies based
on the java version. They do not work with gradle toolchains
and artifacts were published with the new guice already. Will
stick with that for now unless it creates problems. This first
happened in 1.3.6 which was released in early August 2022.
No issues have been reported thus far.