batch updater for counter, timer, and dist summary (#1003)
Adds a helper that can be used to batch updates to the
delegate meter. The main use-case for this is timers or
distributions summaries that need to be updated from a
single thread with a high throughput.

For counters, this can easily be achieved by the user
with a local variable to track the count and then calling
increment with the amount when complete or at the desired
batch interval. However, for timers and distribution
summaries that doesn't work because the implementation
needs the raw samples. In #941 a method was added so the
samples could be accumulated in an array and posted in
a batch. This helps a bit, but it is cumbersome to use,
incurs memory overhead, and the full set of samples have
to be traversed and processed when recorded.

This change introduces batch updaters that can be
customized for a given implementation. So for registries
like Atlas where it is possible to accumulate the stats
without the memory overhead, it can provide an optimized
implementation to do so.