use custom replace for substitutions
The variable substitution when processing JMX data used
the `String.replace` function. On java 8 this internally
uses regex which is costly to compile. This change switches
it to a custom function for replacing the values. On newer
java versions the default is more efficient and doesn't use
regex, but the custom version still seems to be a little
better for the sample data tested.