improve performance of substitution for jmx vars
Before it would loop across the set of variables and try
to perform 2 substitutions for each (raw and decapitalized).
This results in many unnecessary passes and the on java 8
the `String.replace` call uses regex internally which is
quite expensive to compile.

It will now traverse the pattern and lookup the variables
found in the pattern string to perform the substituion. This
means the substitution performance shouldn't vary based on
the number of variables and it avoids the regex issue.