support more customization for evaluator
For some internal uses under heavy load, such as aggregators, it
is desirable to have more control over some of the settings used
with the LWC evaluator. In particular, this allows for tuning
the cache behavior used with the index. Later it can be extended
to support parallelism for matching against the index.