reduce overhead for cached stat id
Updates how the cached stat id used for measurements is
computed on simple meter types. Before it would always
create a new id and then add back the original tags to
overwrite. Now it will check to see if they are present
and only add if missing. This ensures that at most a new
id will be created once and in some cases not at all.