support matching without structuring as an id (#1052)
Add `forEachMatch` overloads to the `QueryIndex` that takes
a function to supply the value for a tag key. For use-cases
where the tags are not already structured as an id, this
avoids the overhead of converting. Since it is just a simple
function, there is a lot of flexibility for the actual data
structure used.