add utility for computing 64-bit hash
Currently based on XXH64. Allows for data to be incrementally
fed to the hash accumulator to make it more convenient to use
and avoid allocations creating the structure needed to pass
into the hash. Though in some cases the incremental processing
is slower than just passing in something like a precomputed
string.