atlas: support delayed gauge aggregation
Add support for delaying gauge aggregation when pushing
data to LWC. In some cases, the registry will be used
on an intermediate cluster that aggregates results and
can receive updates more frequently than the consumption
step size. This can lead to the same gauge value coming
from multiple aggregators leading to skewed values for
accumulating aggregation types like sum and count.

To avoid this problem, the gauge aggregation can be
delayed until the evaluator consuming the data so that
the corresponding values can be deduped before computing
the final aggregation. The deduping is based on the
`atlas.aggr` dimension which is a hash of the id.