determine if pattern is starts-with or contains
Adds support to PatternMatcher to check if a pattern is
a strict starts-with or contains check. This can be useful
when mapping to some data stores that have optimizations
for those operations that will not work with arbitrary
regex. Also adds a method to get a contained string that
can be used as an initial filter.