atlas: fix duplicate flush on shutdown
If it was in the process of sending data when the registry
is shutdown, then it could result in the same time interval
getting sent multiple times. Synchronizing the methods is
a simple solution for now as only a single thread should
be performing a send at a given time.