extract trigrams for patterns (#1073)
Add method to PatternMatcher that can be used to extract
a set of trigrams based on the pattern. For the pattern
to match, a string must contain all of the trigrams. This
only indicates a string might match, the full pattern will
still need to be evaluated on the candidates that contain
all of the trigrams.