api: fix duplicate tag bug in ArrayTagSet (#1074)
It was possible to provide a list of `Tag`s to an `ArrayTagSet` with duplicate
values. The duplicates would not be de-duped at merge time if the duplicate keys in
the list were lexically less than that of existing tags. Merge now checks the last
written tag value to make sure we don't write a duplicate key.

---------

Co-authored-by: Brian Harrington <brharrington@gmail.com>