atlas: ignore unknown properties for lwc subs (#1076)
Update the JSON parsing for the LWC subscription manager
so that it will ignore unknown fields. This avoids breaking
if the server adds some additional metadata to the response.
There is the expectation, that new fields can be added as
long as it doesn't otherwise break the expected behavior for
old clients.