ipc: reset disable flag when reusing entries
Before if there was a disabled request that was successful,
then all subsequent requests that reuse the entry would also
be disabled. Also moves the finalization of some of the fields
to a helper that is always called to avoid different values
going to the access log if metrics are disabled.