Avoid allocating lambda in PercentileTimer.computeIfAbsent (#1128)
PercentileTimer.computeIfAbsent would delegate to Utils.computeIfAbsent, which actually
eagerly calls get and then putIfAbsent if the timer is not in the underlying map. Just directly
do that to avoid needing to allocate a lambda in the case the timer is present. The same
change has been applied to PercentileDistributionSummary.

---------

Co-authored-by: Brian Harrington <brharrington@gmail.com>