reduce memory overhead for step value types (#1158)
Use AtomicLongFieldUpdater to shrink the footprint of StepLong, StepDouble,
and AtomicDouble. This avoids the object overhead and pointer traversal
needed when using AtomicLong internally. 

Note, it does change some of the public methods on StepLong and StepDouble.
However, those are internal implementation classes and should not be used
outside of the spectator libraries.

---------

Co-authored-by: Brian Harrington <brharrington@gmail.com>