atlas: handle OR with multiple matches (#1184)
Update query index to automatically dedup results to handle
the case of OR clauses where both sides match the same data
point. Before, the same match would be returned for each OR
clause that matched a given data point. If the consumer did
not handle that case, then they could aggregate the same
value multiple times leading to incorrect results.

This will add some overhead as the set needs to be maintained,
but should improve correctness for general usage.