add TagListBuilder helper
Helper to support more efficient building of tag lists
with fewer allocations than relying on varargs `withTags`
method. If the builder is reused, then there would only
be a single allocation to copy the buffer array for most
uses.

The user would be responsible for managing the reuse of
the builder instances. This could be a thread local,
concurrent queue, or other options that make sense for
their use-case.